Move platform-level tests to platform/config.rs
Moved 36 platform-level tests from nes/console/config.rs to
platform/config.rs test module:
- Help flag tests (--help, -h, help text structure)
- Audio/video/input CLI tests (audio, vsync, gamepads, fullscreen, display)
- Tracing CLI tests (trace-cpu, trace-ppu, trace-apu, trace-mapper, levels)
- Window height tests
- Debugger enabled test
- Unknown argument handling
- Positional ROM path argument
- Multiple flags integration test
- Breakpoint tests (pc, cycle, frame, write address)
- Autorun tests (playback, trim, convert, recalculate, format)
- TUI mode tests

Config file tests that use apply_config_value remain in nes module
since that method is private and config file parsing is orchestrated
there. These CLI tests verify platform flags work correctly.

Result:
- Platform config tests: 36 (was 0)
- NES config tests: 320 (was 339)
- Total tests: 8680 passing
